refactor(tui): extract skill command handlers component
